move index tests to separate file
